refactor(front): move alerts to separate file
